--- LightningFasterRCNN/src/LightningMetrics.py
@@ -1,2 +1,0 @@
-from pytorch_lightning.metrics import Metric
-from pytorch_lightning.metrics.functional import iou as iou_func
